+++ src/pathfinding_backend.h
@@ -60,7 +60,7 @@ struct DynamicCostOverlay;
 /// integer-grid A* with none of them. The values are serialized, so they are
 /// fixed, only append. A new feature also joins the --pathfindingbackend
 /// valid list in clparse.cpp.
-enum PathfindingFeature : uint8_t
+enum PathfindingFeature : uint16_t
 {
 	PF_DIRECTIONAL_BIAS  = 1 << 0,   ///< directional route shaping after the search
 	PF_CORRIDOR_LANES    = 1 << 1,   ///< steer into lanes through detected corridors
